+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.6.0
 + Implement `dropLast(n)`
++ Implement `grouping()` and `groupingBy(fn)`
 + Add support for `zipWith(iterable)` and `zipWith(iterator)`
 + Add support for `interleave(iterable)` and `interleave(iterator)`
 + Add support for `appendLonger()`, `appendArgumentIfLonger()` and `appendSourceIfLonger()` on `interleave()`